refactor: add dom-api barrel export and reduce coupling in board-view & flow-modal (#600)
Create src/utils/dom-api.js barrel that re-exports all symbols from
dom-core, dom-buttons, dom-lists, dom-dialogs, and dom-tabs.

Update board-view.js (3 dom-* imports → 1) and flow-modal.js
(3 dom-* imports → 1) to use the new barrel. file-viewer.js and
tab-manager.js already use their existing barrel exports
(file-viewer-utils.js and tab-manager-utils.js).

Closes #600
